requirements.txt
clip_by_openai==1.1
einops==0.6.1
kornia==0.7.0
langchain==0.0.286
#numpy==1.24.3
numpy==1.22.4
openai==0.27.3
Pillow==9.4.0
#Pillow==10.0.0
#scipy==1.11.2
timm==0.9.2
#torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
yacs==0.1.8